+++ module/test-dialog.js
@@ -285,11 +285,7 @@ export class TestDialog extends HandlebarsApplicationMixin(ApplicationV2) {
     this.test.isOther2 = !!fields.other2Modifier;
     this.test.isOther3 = !!fields.other3Modifier;
 
-    if (this.mode === 'update') {
-
-      this.test.diceroll = null;
-
-    } else {
+    if (this.mode !== 'update') {
 
       // Set DN Descriptor unless actively defending (in which case no DN, but we set to standard to avoid problems down the line)
       if (this.test.testType === 'activeDefense') this.test.DNDescriptor = 'standard';
@@ -392,6 +388,8 @@ export function oneTestTarget(token, applySize) {
 
     case 'threat':
     case 'stormknight':
+      const skills = {};
+
       return {
         type: actor.type,
         id: actor.id,
@@ -405,10 +403,14 @@ export function oneTestTarget(token, applySize) {
         defenseTraits: actor.defenseTraits,
         // then non-vehicle changes
         skills: actor.itemTypes.customSkill.reduce((acc, skill) => {
-          acc[toCamelCase(skill.name)] = { name: skill.name, ...skill.system };
+          acc[toCamelCase(skill.name)] = { value: skill.system.value, baseAttribute: skill.system.baseAttribute };
           return acc;
-        }, { ...actor.system.skills }),
-        attributes: actor.system.attributes,
+        },
+          Object.entries(actor.system.skills).reduce((acc, [skillName, skill]) => {
+            acc[skillName] = { value: skill.value, baseAttribute: skill.baseAttribute }
+            return acc;
+          }, {})),
+        attributes: Object.entries(actor.system.attributes).reduce((acc, [key, attr]) => (acc[key] = attr.value, acc), {}),
         vulnerableModifier: actor.statusModifiers.vulnerable,
         darknessModifier: actor.statusModifiers.darkness,
         isConcentrating: actor.isConcentrating,